Replaced print statements with logging, renamed deprecated files
